Try reintroducing backpressure into Node.downloadBlocks()
